+++ classes/local/hooks/output/before_standard_head_html_generation.php
@@ -30,6 +30,7 @@ class before_standard_head_html_generation {
      * @param \core\hook\output\before_standard_head_html_generation $hook
      */
     public static function callback(\core\hook\output\before_standard_head_html_generation $hook): void {
+        global $CFG;
 
         if (during_initial_install() || isset($CFG->upgraderunning)) {
             // Do nothing during installation or upgrade.