+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## [4.5.1] - 2026-04-17
 ### Fixed
 - Web service missing validate_context() call - detected by https://mdlshield.com
 - Prevent hooks from being executed during upgrade (missing `global $CFG`)